requirements.txt
Flask
Pillow
pypdf
gunicorn
